version 17del11

+++ grails-app/services/comandapp/PedidoService.groovy
@@ -4,7 +4,7 @@ import grails.transaction.Transactional
 
 @Transactional
 class PedidoService {
-	def getById(int id)
+	def getById(long id)
 	{
 		Pedido.findById(id)
 	}
